% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{server}
\usage{
server(input, output)
}
\arguments{
\item{input}{- parameters input by ui.R}

\item{output}{- output to tabs etc}
}
\description{
This Shiny application uses an adapted version of the Titanic passenger dataset on Kaggle to
demonstrate how analyses and predictions can be made part of an interactive Web-based user
experience.

The server function tests for the presence of the dataset file in the current working 
directory and downloads if not there.  It then generates a dummy id for each passenger record 
before calling the following functions:

tfit -
uses rpart to generate the decision tree model for the whole dataset

tpred - 
returns a prediction based on that model for whatever slice of data is sent 
to the function

tfilter - 
reactive function to slice the data based on the curren parameter values

survivors - 
reactive function to slice the data returned by tfilter so that it contains 
only records for survivors

Then the script generates the output for the 3 tabs as follows:

Output for tab 1 - Survival by Age Group 

Updates the headline comparing the number of passengers and overall probability 
of survival for the current dataset with those of the whole dataset

Output for tab 2 - 3D Analysis

Plots two 3D scatter charts side by side.  Each plots survived (green) vs perished (red).
Despite the three dimensions, there is still some over-plotting, hence the opacity.
The first chart analyses the combined influence of the Age, Gender & Family Members predictors.
The second chart analyses the combined influence of the Age, Port of Embarcation & Spouses/Siblings
predictors. 

Outputs for tab 3 - Decision Tree -

Uses the rattle package and the fancyRpartPlot function
to generate a full decision tree.
}
