% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{ui.R}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
ui
}
\arguments{
\item{Sex/Gender}{- female/male}

\item{Class}{- passenger's class of travel - 1, 2 or 3}

\item{Embarked}{- port of embarcation, one of Cherbourg, Queenstown or Southampton}

\item{SibSp}{- Nr of Siblings or Spouses accompanying - 1 to 10}

\item{Parch}{- Nr of Parents or Children accompanying - 1 to 10}
}
\description{
This Shiny application uses an adapted version of the Titanic passenger dataset on Kaggle to
demonstrate how analyses and predictions can be made part of an interactive Web-based user
experience.

The ui function defines the page layout for the Shiny application.  It places a parameter 
panel on the left-hand side leaving space for 3 tabs over the rest of the page.  The initial 
parameter settings and view display the whole dataset.  Users are then free to adjust any of 
the parameters and so slice the current data.  As the view of the data changes, so too will 
the probability of survival and the content of the 3 tabs.

The 3 tabs are:

1 - Histogram of passengers vs survivors

2 - 3D charts showing the influence of the predictor variables on survival

3 - Decision tree of probabilities for the current data slice
}
\keyword{datasets}
